refactor(ingestion): move TopHogComponent next to the TopHog class
