+++ nodejs/src/ingestion/framework/tophog/index.ts
@@ -1,2 +1,2 @@
-export { TopHog } from './tophog'
+export { TopHog, TopHogComponent } from './tophog'
 export type { TopHogRequiredConfig, TopHogOptionalConfig } from './tophog'
